+++ lib/src/credentials.dart
@@ -46,7 +46,9 @@ enum AuthProviderType {
   /// For authenticating with an email and a password.
   emailPassword,
 
-  _function,
+  /// For authenticating with custom function with payload argument.
+  function,
+
   _userApiKey,
   _serverApiKey
 }
@@ -96,7 +98,13 @@ class Credentials {
   /// Returns a [Credentials] object that can be used to authenticate a user with a Google account using an id token.
   Credentials.googleIdToken(String idToken)
       : _handle = realmCore.createAppCredentialsGoogleIdToken(idToken),
-        provider = AuthProviderType.google;
+        provider = AuthProviderType.google;  
+  
+  /// Returns a [Credentials] object that can be used to authenticate a user with a custom Function.
+  /// [Custom Function Authentication Docs](https://www.mongodb.com/docs/atlas/app-services/authentication/custom-function/)
+  Credentials.function(String payload)
+      : _handle = realmCore.createAppCredentialsFunction(payload),
+        provider = AuthProviderType.function;
 }
 
 /// @nodoc